requirements.txt
# AIOps Multi-Agent System Requirements

# Core LangChain dependencies
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0

# LLM providers
openai>=1.0.0

# Tavily AI Search (for real-time information retrieval)
tavily-python>=0.3.0

# Vector store for GRPO knowledge base
chromadb>=0.4.0

# HTTP client for tool integrations
requests>=2.31.0

# Data validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Optional: For async support
# aiohttp>=3.9.0

# Optional: For advanced logging
# structlog>=24.0.0

# Optional: For testing
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
